Implement logo upload and all products view
