+++ pages/staff.js
@@ -390,6 +390,26 @@ export default function StaffPortal() {
                   📸 Upload Images
                 </button>
               )}
+
+              {/* View All Products Button - Only show on inventory tab */}
+              {activeTab === 'inventory' && (
+                <button
+                  onClick={() => router.push('/all-products')}
+                  style={{
+                    background: 'linear-gradient(135deg, #667eea 0%, #764ba2 100%)',
+                    color: 'white',
+                    border: 'none',
+                    padding: '12px 20px',
+                    borderRadius: '6px',
+                    cursor: 'pointer',
+                    fontSize: '14px',
+                    fontWeight: 'bold',
+                    boxShadow: '0 2px 4px rgba(0,0,0,0.2)'
+                  }}
+                >
+                  📋 All Products
+                </button>
+              )}
             </div>
           </div>
         </div>
